+++ cakk-api/src/main/java/com/cakk/api/service/shop/ShopService.java
@@ -19,7 +19,7 @@ import com.cakk.api.dto.response.shop.CakeShopSearchResponse;
 import com.cakk.api.dto.response.shop.CakeShopSimpleResponse;
 import com.cakk.api.mapper.PointMapper;
 import com.cakk.api.mapper.ShopMapper;
-import com.cakk.common.dto.OperationDays;
+import com.cakk.api.dto.request.shop.OperationDays;
 import com.cakk.domain.mysql.bo.CakeShops;
 import com.cakk.domain.mysql.dto.param.cake.CakeImageResponseParam;
 import com.cakk.domain.mysql.dto.param.shop.CakeShopByKeywordParam;
